+++ src/components/Menu/Menu/Menu.jsx
@@ -31,7 +31,8 @@ const Menu = forwardRef(
       focusItemIndex,
       focusItemIndexOnMount,
       isSubMenu,
-      useDocumentEventListeners
+      useDocumentEventListeners,
+      shouldScrollMenu
     },
     forwardedRef
   ) => {
@@ -146,7 +147,8 @@ const Menu = forwardRef(
                   closeMenu: onCloseMenu,
                   menuId: id,
                   useDocumentEventListeners,
-                  isInitialSelectedState
+                  isInitialSelectedState,
+                  shouldScrollMenu
                 })
               : null;
           })}
@@ -172,7 +174,8 @@ Menu.defaultProps = {
   focusItemIndex: -1,
   isSubMenu: false,
   useDocumentEventListeners: false,
-  focusItemIndexOnMount: -1
+  focusItemIndexOnMount: -1,
+  shouldScrollMenu: false
 };
 
 Menu.propTypes = {
@@ -188,7 +191,8 @@ Menu.propTypes = {
   focusItemIndex: PropTypes.number,
   isSubMenu: PropTypes.bool,
   useDocumentEventListeners: PropTypes.bool,
-  focusItemIndexOnMount: PropTypes.number
+  focusItemIndexOnMount: PropTypes.number,
+  shouldScrollMenu: PropTypes.bool
 };
 
 export default Menu;